+++ app/src/main/res/layout/item_cart.xml
@@ -11,7 +11,7 @@
 
         <variable
             name="onClickHandler"
-            type="woowacourse.shopping.ui.cart.CartViewHolder.OnClickHandler" />
+            type="woowacourse.shopping.ui.cart.adapter.CartViewHolder.OnClickHandler" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
